requirements.txt
Pillow
pymongo